feat: endpoints, migrations fix
Implementation of functional endpoints, minor changes to migrations due to type changes (ex: int -> float64 or string -> int)

+++ qualify-backend/pkg/analyst.go
@@ -4,5 +4,5 @@ type Analyst struct {
 	User
 	Hourly_rate   float64 `json:"hourly_rate"`
 	Total_reviews int     `json:"total_reviews"`
-	Mean_rating   int     `json:"mean_rating"`
+	Mean_rating   float64 `json:"mean_rating"`
 }
